cleanup: move workload finish to a subpackage
