+++ pkg/controller/tas/indexer/indexer.go
@@ -28,6 +28,7 @@ import (
 	utiltas "sigs.k8s.io/kueue/pkg/util/tas"
 	"sigs.k8s.io/kueue/pkg/workload"
 	workloadevict "sigs.k8s.io/kueue/pkg/workload/evict"
+	workloadfinish "sigs.k8s.io/kueue/pkg/workload/finish"
 )
 
 const (
@@ -90,7 +91,7 @@ func indexResourceFlavorTopologyName(o client.Object) []string {
 
 func indexAdmittedWorkloadNodes(o client.Object) []string {
 	wl, ok := o.(*kueue.Workload)
-	if !ok || workload.IsFinished(wl) || workloadevict.IsEvicted(wl) {
+	if !ok || workloadfinish.IsFinished(wl) || workloadevict.IsEvicted(wl) {
 		return nil
 	}
