+++ pkg/workload/workload_test.go
@@ -2354,110 +2354,6 @@ func TestWorkloadPriorityClassChanged(t *testing.T) {
 	}
 }
 
-func TestFinish(t *testing.T) {
-	const (
-		baseReason  = "TestReason"
-		baseMessage = "Test Message"
-	)
-
-	now := time.Now().Truncate(time.Second)
-
-	baseWl := utiltestingapi.MakeWorkload("wl", metav1.NamespaceDefault).ResourceVersion("1")
-
-	type args struct {
-		wl       *kueue.Workload
-		reason   string
-		message  string
-		patchErr error
-	}
-
-	type want struct {
-		wl  *kueue.Workload
-		err error
-	}
-
-	tests := map[string]struct {
-		args args
-		want want
-	}{
-		"finish workload": {
-			args: args{
-				wl:      baseWl.DeepCopy(),
-				reason:  baseReason,
-				message: baseMessage,
-			},
-			want: want{
-				wl: baseWl.Clone().
-					ResourceVersion("2").
-					Condition(metav1.Condition{
-						Type:               kueue.WorkloadFinished,
-						Status:             metav1.ConditionTrue,
-						LastTransitionTime: metav1.NewTime(now),
-						Reason:             baseReason,
-						Message:            baseMessage,
-					}).
-					Obj(),
-			},
-		},
-		"already finished workload": {
-			args: args{
-				wl:      baseWl.Clone().FinishedAt(now).Obj(),
-				reason:  "OtherReason",
-				message: "Other Message",
-			},
-			want: want{
-				wl: baseWl.Clone().FinishedAt(now).Obj(),
-			},
-		},
-		"error on finish": {
-			args: args{
-				wl:       baseWl.DeepCopy(),
-				reason:   baseReason,
-				message:  baseMessage,
-				patchErr: errTest,
-			},
-			want: want{
-				wl:  baseWl.DeepCopy(),
-				err: errTest,
-			},
-		},
-	}
-	for name, tc := range tests {
-		t.Run(name, func(t *testing.T) {
-			ctx, _ := utiltesting.ContextWithLog(t)
-
-			cl := utiltesting.NewClientBuilder().
-				WithObjects(tc.args.wl).
-				WithStatusSubresource(&kueue.Workload{}).
-				WithInterceptorFuncs(interceptor.Funcs{
-					SubResourcePatch: func(ctx context.Context, c client.Client, subResourceName string, obj client.Object, patch client.Patch, opts ...client.SubResourcePatchOption) error {
-						if tc.args.patchErr != nil {
-							return tc.args.patchErr
-						}
-						return utiltesting.TreatSSAAsStrategicMerge(ctx, c, subResourceName, obj, patch, opts...)
-					},
-				}).
-				Build()
-
-			fakeClock := testingclock.NewFakeClock(now)
-
-			gotErr := Finish(ctx, cl, tc.args.wl, tc.args.reason, tc.args.message, fakeClock)
-			if diff := cmp.Diff(tc.want.err, gotErr, cmpopts.EquateErrors()); diff != "" {
-				t.Errorf("Unexpected error (-want,+got):\n%s", diff)
-			}
-
-			updatedWl := &kueue.Workload{}
-			if err := cl.Get(ctx, client.ObjectKeyFromObject(tc.args.wl), updatedWl); err != nil {
-				t.Fatalf("Failed obtaining updated object: %v", err)
-			}
-
-			if diff := cmp.Diff(tc.want.wl, updatedWl, cmpopts.EquateEmpty()); diff != "" {
-				t.Errorf("Unexpected workload (-want,+got):\n%s", diff)
-			}
-		})
-	}
-}
-
 func TestEvictionPendingLatency(t *testing.T) {
 	evictTime := time.Date(2024, 3, 15, 12, 0, 0, 0, time.UTC)
 	metricNow := evictTime.Add(30 * time.Second)
