+++ test/e2e/singlecluster/extended/kuberay_test.go
@@ -45,6 +45,7 @@ import (
 	testingrayjob "sigs.k8s.io/kueue/pkg/util/testingjobs/rayjob"
 	testingrayservice "sigs.k8s.io/kueue/pkg/util/testingjobs/rayservice"
 	"sigs.k8s.io/kueue/pkg/workload"
+	workloadfinish "sigs.k8s.io/kueue/pkg/workload/finish"
 	"sigs.k8s.io/kueue/pkg/workloadslicing"
 	"sigs.k8s.io/kueue/test/util"
 )
@@ -341,7 +342,7 @@ print(ray.get([my_task.remote(i, 10) for i in range(20)]))`,
 				g.Expect(workloadList.Items).NotTo(gomega.BeEmpty(), "Expected at least one workload in namespace")
 				hasAdmittedWorkload := false
 				for _, wl := range workloadList.Items {
-					if workload.IsAdmitted(&wl) || workload.IsFinished(&wl) {
+					if workload.IsAdmitted(&wl) || workloadfinish.IsFinished(&wl) {
 						hasAdmittedWorkload = true
 						break
 					}
@@ -541,7 +542,7 @@ print([ray.get(my_task.remote(i, 1)) for i in range(20)])`,
 				g.Expect(workloadList.Items).NotTo(gomega.BeEmpty(), "Expected at least one workload in namespace")
 				hasAdmittedWorkload := false
 				for _, wl := range workloadList.Items {
-					if workload.IsAdmitted(&wl) || workload.IsFinished(&wl) {
+					if workload.IsAdmitted(&wl) || workloadfinish.IsFinished(&wl) {
 						hasAdmittedWorkload = true
 						break
 					}
@@ -1059,7 +1060,7 @@ app = HelloWorld.bind()`,
 				g.Expect(k8sClient.List(ctx, workloadList, client.InNamespace(ns.Name))).To(gomega.Succeed())
 				var activeWorkloads []kueue.Workload
 				for i := range workloadList.Items {
-					if !workload.IsFinished(&workloadList.Items[i]) {
+					if !workloadfinish.IsFinished(&workloadList.Items[i]) {
 						activeWorkloads = append(activeWorkloads, workloadList.Items[i])
 					}
 				}
