+++ test/performance/scheduler/runner/recorder/recorder.go
@@ -32,6 +32,7 @@ import (
 	kueue "sigs.k8s.io/kueue/apis/kueue/v1beta2"
 	"sigs.k8s.io/kueue/pkg/workload"
 	workloadevict "sigs.k8s.io/kueue/pkg/workload/evict"
+	workloadfinish "sigs.k8s.io/kueue/pkg/workload/finish"
 	"sigs.k8s.io/kueue/test/performance/scheduler/runner/generator"
 )
 
@@ -460,7 +461,7 @@ func (r *Recorder) RecordWorkloadState(wl *kueue.Workload) {
 		ClassName: wl.Labels[generator.ClassLabel],
 		Admitted:  workload.IsAdmitted(wl),
 		Evicted:   workloadevict.IsEvicted(wl),
-		Finished:  workload.IsFinished(wl),
+		Finished:  workloadfinish.IsFinished(wl),
 	}
 	select {
 	case r.wlEvChan <- ev:
